+++ mo-server/src/main/java/com/mo/service/DashBoardService.java
@@ -10,13 +10,14 @@ public interface DashBoardService {
      * 获取总文章阅读量
      * @return
      */
-    Long total();
+    Integer total();
 
     /**
      * 获取今天阅读量
+     *
      * @return
      */
-    Long todayView();
+    Integer todayView();
 
     /**
      * 获取未读评论
